image_generator/requirements.txt
redis
Pillow
rq